Extract authentication handlers classes.
