+++ s3/src/main/java/ch/cyberduck/core/s3/S3Session.java
@@ -22,7 +22,6 @@ package ch.cyberduck.core.s3;
 import ch.cyberduck.core.AttributedList;
 import ch.cyberduck.core.Credentials;
 import ch.cyberduck.core.CredentialsConfigurator;
-import ch.cyberduck.core.DisabledTranscriptListener;
 import ch.cyberduck.core.Host;
 import ch.cyberduck.core.HostKeyCallback;
 import ch.cyberduck.core.ListProgressListener;
@@ -77,7 +76,6 @@ import ch.cyberduck.core.threading.CancelCallback;
 import ch.cyberduck.core.transfer.TransferStatus;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.http.Header;
 import org.apache.http.HttpHeaders;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpRequest;
@@ -85,7 +83,6 @@ import org.apache.http.HttpRequestInterceptor;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpPut;
-import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.protocol.HTTP;
 import org.apache.http.protocol.HttpContext;
@@ -99,23 +96,12 @@ import org.jets3t.service.model.StorageObject;
 import org.jets3t.service.security.AWSCredentials;
 import org.jets3t.service.security.AWSSessionCredentials;
 import org.jets3t.service.security.ProviderCredentials;
-import org.jets3t.service.utils.RestUtils;
-import org.jets3t.service.utils.ServiceUtils;
 import org.jets3t.service.utils.SignatureUtils;
 
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
-
-import com.amazonaws.auth.internal.SignerConstants;
 
 import static com.amazonaws.services.s3.Headers.*;
 
@@ -293,136 +279,14 @@ public class S3Session extends HttpSession<RequestEntityRestStorageService> {
                                     new AWSProfileSTSCredentialsConfigurator(new ThreadLocalHostnameDelegatingTrustManager(trust, host.getHostname()), key, prompt) : CredentialsConfigurator.DISABLED
             ));
         }
-        configuration.addInterceptorLast(new HttpRequestInterceptor() {
-            @Override
-            public void process(final HttpRequest request, final HttpContext context) throws IOException {
-                if(!client.isAuthenticatedConnection()) {
-                    log.warn(String.format("Skip authentication request %s", request));
-                    return;
-                }
-                final ProviderCredentials credentials = client.getProviderCredentials();
-                final String bucketName = context.getAttribute("bucket").toString();
-                if(log.isDebugEnabled()) {
-                    log.debug(String.format("Use bucket name %s from context for %s signature", bucketName, authenticationHeaderSignatureVersion));
-                }
-                final URI uri;
-                try {
-                    uri = new URI(request.getRequestLine().getUri());
-                }
-                catch(URISyntaxException e) {
-                    throw new IOException(e);
-                }
-                switch(authenticationHeaderSignatureVersion) {
-                    case AWS2: {
-                        String path = uri.getRawPath();
-                        // If bucket name is not already part of the full path, add it.
-                        // This can be the case if the Host name has a bucket-name prefix,
-                        // or if the Host name constitutes the bucket name for DNS-redirects.
-                        if(!StringUtils.startsWith(path, bucketName)) {
-                            path = String.format("/%s%s", bucketName, path);
-                        }
-                        final String queryString = uri.getRawQuery();
-                        if(StringUtils.isNotBlank(queryString)) {
-                            path += String.format("?%s", queryString);
-                        }
-                        // Generate a canonical string representing the operation.
-                        final String canonicalString = RestUtils.makeServiceCanonicalString(
-                                request.getRequestLine().getMethod(),
-                                path,
-                                this.getHeadersAsObject(request),
-                                null,
-                                getRestHeaderPrefix(),
-                                client.getResourceParameterNames());
-                        // Sign the canonical string.
-                        final String signedCanonical = ServiceUtils.signWithHmacSha1(
-                                credentials.getSecretKey(), canonicalString);
-                        // Add encoded authorization to connection as HTTP Authorization header.
-                        final String authorizationString = getSignatureIdentifier() + " "
-                                + credentials.getAccessKey() + ":" + signedCanonical;
-                        request.setHeader(HttpHeaders.AUTHORIZATION, authorizationString);
-                        break;
-                    }
-                    case AWS4HMACSHA256: {
-                        String region = regions.getRegionForBucketName(bucketName);
-                        if(null == region) {
-                            final HttpHost host = (HttpHost) context.getAttribute(HttpCoreContext.HTTP_TARGET_HOST);
-                            if(host != null) {
-                                try {
-                                    region = SignatureUtils.awsRegionForRequest(new URI(host.toURI()));
-                                }
-                                catch(URISyntaxException e) {
-                                    throw new IOException(e);
-                                }
-                            }
-                            if(region != null) {
-                                if(log.isDebugEnabled()) {
-                                    log.debug(String.format("Cache region %s for bucket %s", region, bucketName));
-                                }
-                                regions.putRegionForBucketName(bucketName, region);
-                            }
-                        }
-                        if(null == region) {
-                            region = host.getRegion();
-                        }
-                        if(null == region) {
-                            region = new HostPreferences(host).getProperty("s3.location");
-                        }
-                        final HttpUriRequest message = (HttpUriRequest) request;
-                        String requestPayloadHexSHA256Hash =
-                                SignatureUtils.awsV4GetOrCalculatePayloadHash(message);
-                        message.setHeader(SignerConstants.X_AMZ_CONTENT_SHA256, requestPayloadHexSHA256Hash);
-                        // Generate AWS-flavoured ISO8601 timestamp string
-                        final String timestampISO8601 = message.getFirstHeader(S3_ALTERNATE_DATE).getValue();
-                        // Canonical request string
-                        final String canonicalRequestString =
-                                SignatureUtils.awsV4BuildCanonicalRequestString(uri,
-                                        request.getRequestLine().getMethod(), this.getHeadersAsString(request), requestPayloadHexSHA256Hash);
-                        // String to sign
-                        final String stringToSign = SignatureUtils.awsV4BuildStringToSign(
-                                authenticationHeaderSignatureVersion.toString(), canonicalRequestString,
-                                timestampISO8601, region);
-                        // Signing key
-                        final byte[] signingKey = SignatureUtils.awsV4BuildSigningKey(
-                                credentials.getSecretKey(), timestampISO8601, region);
-                        // Request signature
-                        final String signature = ServiceUtils.toHex(ServiceUtils.hmacSHA256(
-                                signingKey, ServiceUtils.stringToBytes(stringToSign)));
-                        // Authorization header value
-                        final String authorizationHeaderValue =
-                                SignatureUtils.awsV4BuildAuthorizationHeaderValue(
-                                        credentials.getAccessKey(), signature,
-                                        authenticationHeaderSignatureVersion.toString(), canonicalRequestString,
-                                        timestampISO8601, region);
-                        message.setHeader(HttpHeaders.AUTHORIZATION, authorizationHeaderValue);
-                        break;
-                    }
-                }
-            }
-
-            final class HttpHeaderFilter implements Predicate<Header> {
-                @Override
-                public boolean test(final Header header) {
-                    return !new HostPreferences(host).getList("s3.signature.headers.exclude").stream()
-                            .filter(s -> StringUtils.equalsIgnoreCase(s, header.getName())).findAny().isPresent();
-                }
-            }
-
-            private Map<String, String> getHeadersAsString(final HttpRequest request) {
-                final Map<String, String> headers = new HashMap<>();
-                for(Header header : Arrays.stream(request.getAllHeaders()).filter(new HttpHeaderFilter()).collect(Collectors.toList())) {
-                    headers.put(StringUtils.lowerCase(StringUtils.trim(header.getName())), StringUtils.trim(header.getValue()));
-                }
-                return headers;
-            }
-
-            private Map<String, Object> getHeadersAsObject(final HttpRequest request) {
-                final Map<String, Object> headers = new HashMap<>();
-                for(Header header : Arrays.stream(request.getAllHeaders()).filter(new HttpHeaderFilter()).collect(Collectors.toList())) {
-                    headers.put(StringUtils.lowerCase(StringUtils.trim(header.getName())), StringUtils.trim(header.getValue()));
-                }
-                return headers;
-            }
-        });
+        switch(authenticationHeaderSignatureVersion) {
+            case AWS4HMACSHA256:
+                configuration.addInterceptorLast(new S3AWS4SignatureRequestInterceptor(this));
+                break;
+            case AWS2:
+                configuration.addInterceptorLast(new S3AWS2SignatureRequestInterceptor(this));
+                break;
+        }
         final RequestEntityRestStorageService client = new RequestEntityRestStorageService(this, configuration);
         client.setRegionEndpointCache(regions);
         return client;
